Added failing test: Apply_Changes_Should_Mark_Unchanged_Order_Unchanged_Customer_With_Multiple_Addresses_Added.  DbContext.ApplyChanges throws the following exception:
Attaching an entity of type 'TrackableEntities.EF.Tests.NorthwindModels.CustomerAddress' failed because another entity of the same type already has the same primary key value. This can happen when using the 'Attach' method or setting the state of an entity to 'Unchanged' or 'Modified' if any entities in the graph have conflicting key values. This may be because some entities are new and have not yet received database-generated key values. In this case use the 'Add' method or the 'Added' entity state to track the graph and then set the state of non-new entities to 'Unchanged' or 'Modified' as appropriate.